distgit: don't check that specfile matches the repository name
Fix #3365

The rpkg-1.67-1 first introduced this check which was implemented in
https://pagure.io/rpkg/pull-request/700

At first I thought the check is valid and we only have incorrectly
named spec files and repositories in our tests. But rpkg-1.67-1
actually breaks importing even for our real usage. This leads me to
think that when they compare the spec and repo name, they forget to
strip a namespace from the repo name.