rpmbuild: make_srpmbuild, set correctly recursive safe.directory
We thought in 77dbe365 that safe.directory is that one what is
specified and everything under. Hovewer this is not true, safe directory
is considered only the mentioned and anything below is excluded.

To consider directory A and anything under safe, gitconfig should look like
this:
[safe]
directory = <path to A>
directory = <path to A>/*

Check out this commit in upstream git that implements this feature:
https://github.com/git/git/commit/313eec177ad010048b399d6fd14de871b517f7e3