backend: fix createrepo action for Pulp CoprDirs
This method is called when backend receives a createrepo action. As you can see,
the `dirname` parameter was unused. Which worked fine when the createrepo action
was generated when creating a new project or when manually clicking the
"Regenerate" button in a project overview. However, createrepo action is also
emitted when somebody submits a build into a CoprDir and thereby creates it.