backend: migrate os.listdir() to os.scandir() to increase performance
As recommended in https://peps.python.org/pep-0471/

scandir() has significantly better performance. On some places it does
not have sense, but I was doing it everywhere, because it was super
easy. I done it, because I recently done the same thing in
scancode-toolkit where it matters a lot.

Related: #3601