frontend: parse every error message to utf-8, ignore non-parsable chars
If user sends to API some weird non-utf-8 character and we then
try to log this, the logging freaks out with `logging error` (see [1]).
[2] fixes this, but only for one endpoint and one type of user input,
but there are many ways how user can do this. On the other hand, [2]
caused bad formating and now we got messages like `b'Project a/b does
not exist'`.

This patch tries to parse every error message we want to store, and maybe
later log it, to utf-8 and store it as a string. If some non-utf-8
character appears in the message output, its byte won't be pares and it
will be stored/printed as the string representation of that byte.

Example:

User calls https://copr.fedorainfracloud.org/coprs/g/some<nasty_characters>x02string

the string value in CoprHttpException.message will be some\x01\x02string

where the \x01\x02 is example of the <nasty_characters>

[1] - https://github.com/fedora-copr/copr/issues/3072
[2] - https://github.com/fedora-copr/copr/pull/3076

Fix #3644