frontend: fix local builds in Pulp projects
Close #3632

In Fedora Copr production, we will leave this value as `None` because we
want people to use our CDN URLs and basically not even notice, the results moved
to Pulp. We already have Lighttpd redirects and everything.

However, in docker-compose, we need to use an URL that is accessible by other
containers.