Adds batching to add_content() and delete_content() in Pulp client.
The locking logic is borrowed from the implementation in the BackendStorage.

Each call to PulpClient.add_content() and PulpClient.delete_content() now works like this:

1. Writes the request to Redis.
2. Check if another process processed the request. Return if so.
2. Try to acquire a lock.
3. If lock is acquired
       a. Collect all requests to add and remove content and make one request to Pulp.
       b. Create a publication
   If lock is not acquire, go back to step 2.
4. Notify other processes that their request was processed.

The PulpStorage.publish_repository() is now a no-op.

The distribution is no longer updated after a publication is created. The distribution always points to the
latest publication for a particular repository.

This reduces the number of new repository versions and publications that are created for large repositories.