frontend: redirect to the current URL after OIDC login
See #3788
See #3769

This uses session to remember the current URL before we get redirected to OIDC
and then destroys it after a redirect. This is the only way (confirmed by
@abompard) because an OIDC provider can only redirect to a whitelisted URL,
which cannot contain a wildcard.
So it cannot be https://copr.fedorainfracloud.org/oidc_auth?next=*
Session variable is also what Bodhi does.