backend, frontend, common: specify reason for running createrepo
We run createrepo in many different scenarios, see
https://docs.pagure.org/copr.copr/createrepo.html#when-do-we-run-createrepo

In different situations we may want to run it differently. That's why I am
introducing the `CreaterepoReason` enum which allows frontned to specify why
the createrepo is being run, and backend can do any special cases based on that.
The corner case that forced me to implement this was a special case when
disabling a manual createrepo feature for a project. I wasn't able to
distinguish that case based on the already available information.

This could also be a first step towards codifying the "When do we run
createrepo" table.