frontend, backend, python, cli: drop support for building modules
As of October 2024, all Modularity features in Copr, except for module
hotfixes and installing modules to buildroot, have been deprecated and
are planned for removal from the codebase. This decision aligns with
the broader retirement of the Fedora Modularity project, which
concluded with Fedora 39 and RHEL 9.

For more details on this transition, please see:
https://frostyx.cz/posts/copr-modularity-the-end-of-an-era