backend: add a helper script to query Pulp results
Sometimes we need to query information from Pulp. That can of course be done
using the `pulp` CLI tool. For example:

    pulp rpm repository show --name @copr/copr/fedora-rawhide-x86_64
    pulp rpm distribution show --name @copr/copr/fedora-rawhide-x86_64

But in certain scenarios we use custom queries that are not easily writeable on
the command line, and even if so, we would like to see what results get queried
by our actual code, not its CLI alternative.

An example of a custom query would be querying RPM results based on a build ID