backend: fix duplicate NEVRA error when migrating projects to Pulp
While trying to migrate the `@copr/copr-pull-requests` project to Pulp, we
encountered the following error:

    Cannot create repository version. More than one rpm.package content with the
    duplicate values for name, epoch, version, release, arch, location_href.

This is a common Copr use-case, building a package with the same NEVRA but
different content again and again, so why the error started happening only now?

Everything works perfectly when submitting consequent builds with same NEVRA and
different content because createrepo happens after every build and therefore
Pulp knows it is supposed to remove the old package from the repository and add
the new one.

However, when migrating data using the `copr-change-storage` scrpit, we upload
all the RPM packages and then do createrepo only once for all of them. But when
we want to add multiple packages at the same time, Pulp doesn't know which of
them is the old one.

In this PR I am implementing a mechanism that prevents migrating duplicate
NEVRAs into a Pulp repository.