testing-farm: install pre-release packages with Ansible
It is better to keep the test script focused solely on testing, separate
from machine preparation.  I needed to revamp some old scripting used
for GitHub automation to make it compatible with Testing Farm as well.

I also switched the logic to work with pull-request "head" references,
which remain stable (unlike "merge" pull-request references, which are
updated when other PRs are merged).