testing-farm: install pre-release packages with Ansible
It's better to keep the test script focused on testing, not machine
preparation.  I needed to revamp some old scripting related to GitHub
automation so it is usable for Testing Farm, too.

I also switched the logic to work with pull-request "head" references
that remain stable ("merge" pull-request refs are updated with other PRs
merged).